+++ android/surge/src/main/java/co/instil/surge/client/SurgeVideoScale.java
@@ -1,0 +1,16 @@
+/*
+ * Copyright (c) 2020 Instil Software.
+ *
+ * This file is subject to the terms and conditions defined in
+ * file 'LICENSE.txt', which is part of this source code package.
+ */
+
+package co.instil.surge.client;
+
+/**
+ * Enum containing the options available for Surge video scaling.
+ */
+public enum SurgeVideoScale {
+    MATCH_ASPECT_RATIO,
+    STRETCH
+}