+++ android/surge/src/test/java/co/instil/surge/decoders/h264/StubH264Decoder.java
@@ -8,13 +8,14 @@
 package co.instil.surge.decoders.h264;
 
 import android.media.MediaCodec;
-import android.view.Surface;
-import co.instil.surge.decoders.MediaCodecFactory;
-import co.instil.surge.decoders.h264.nalu.NaluParser;
-import co.instil.surge.device.DeviceExaminer;
 
 import java.util.ArrayList;
 import java.util.List;
+
+import co.instil.surge.client.SurgeVideoView;
+import co.instil.surge.decoders.MediaCodecFactory;
+import co.instil.surge.decoders.h264.nalu.NaluParser;
+import co.instil.surge.device.DeviceExaminer;
 
 /**
  * Created by mattmccomb on 08/07/16.
@@ -23,17 +24,17 @@ public class StubH264Decoder extends H264Decoder {
 
     private List<H264Packet> receivedPackets;
 
-    protected StubH264Decoder(Surface surface) {
-        super(surface);
+    protected StubH264Decoder(SurgeVideoView videoView) {
+        super(videoView);
         receivedPackets = new ArrayList<>();
     }
 
     protected StubH264Decoder(
-            Surface surface,
+            SurgeVideoView videoView,
             MediaCodecFactory mediaCodecFactory,
             NaluParser naluParser,
             DeviceExaminer deviceExaminer) {
-        super(surface, mediaCodecFactory, naluParser, deviceExaminer);
+        super(videoView, mediaCodecFactory, naluParser, deviceExaminer);
         receivedPackets = new ArrayList<>();
     }