+++ xamarin/build-android.sh
@@ -8,7 +8,7 @@ rm $BUILD_DIR/SurgeAndroid.aar
 
 cd $SURGE_DIR
 
-./gradlew surge:build -x checkstyle -x findbugs
+./gradlew surge:build
 
 cp $SURGE_DIR/surge/build/outputs/aar/surge-release.aar $BUILD_DIR/SurgeAndroid.aar